Improve SearchTaskData struct, do not use JSON directly, move structs to structs.go
